SM1 - Finalized Dynamic Tablet Login Screen UI
* Modified Login Screen by making it dynamic to accomodate wider widths

* Modified Registration Screen by correcting the name of the Large Body Widget

* Modified Background_Image.psd by adding a seperate design for the Login Screen

* Modified pubscpec.yaml by adding the Login Screen Image

* Added Background_image_login.jpg to assets
