+++ pubspec.yaml
@@ -87,6 +87,7 @@ flutter:
     - assets\images\JPG\doctor_three.jpg
     - assets\images\JPG\RX.jpg
     - assets\images\JPG\background_image_sugbodoc.jpg
+    - assets\images\JPG\background_image_login.jpg
